requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
bottle==0.13.4
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
clr_loader==0.2.7.post0
colorama==0.4.6
ConfigArgParse==1.7.1
distlib==0.3.9
djando==0.2
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
facebook-sdk==3.1.0
Faker==20.1.0
fastapi==0.115.4
filelock==3.16.1
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.2
frozenlist==1.5.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.17.0
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.2.0
mysql-connector-python==9.1.0
mysqlclient==2.2.7
packaging==25.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
propcache==0.3.1
proxy_tools==0.1.0
psutil==7.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==7.4.3
pytest-html==4.1.1
pytest-metadata==3.1.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-escpos==3.1
python-http-client==3.3.7
pythonnet==3.0.5
pytz==2025.1
pyusb==1.3.1
pywebview==5.4
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.0
qrcode==8.0
razorpay==1.4.2
reportlab==4.4.2
requests==2.31.0
roundrobin==0.0.4
sendgrid==3.6.5
sendgrid-django==4.2.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.2
twilio==9.5.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.32.0
virtualenv==20.27.1
Werkzeug==3.0.1
WTForms==3.2.1
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
